+++ RELEASE_NOTES.md
@@ -5,8 +5,19 @@ A new header is inserted each time a *tag* is created.
 
 ## Next release
 
+- Fixed morph shapes not rendering in WebGL.
+- Added support for the latest version of emscripten.
+- gltfio: fixed blackness seen with default material.
 - Added ETC2 and BC compressed texture support to Metal backend.
-- Add Java / Kotlin bindings for gltfio::Animator.
+
+## v1.4.2
+
+- Cleaned up the validation strategy in Engine (checks for use-after-destroy etc).
+- OpenGL: Fixed ES 3.0 support on iOS.
+- OpenGL: Added support for KHR_debug in debug builds.
+- gltfio: Added Java / Kotlin bindings for Animator.
+- gltfio: Fixed panic with the Android gltf-bloom demo.
+- gltfio: Java clients should no longer call Filament#init.
 
 ## v1.4.1